[Project, ECS] Preventing flaw with templated GameObjects/Entitiies causing potential slicing of the handles. # - Removed functionality of templated GameObjects/Entitiies. Entities cannot be inherited from and GameObjects are just aliasses. - EntityTypeID now uses FamilyType based on provided type names instead of type classes. - Added FamilyNameType which is essentially the same as the FamilyClassType to provide quick ID for classes, but now for a given family name/string.

+++ Digital/Digital_/Source/GameWorld/GameWorld.cpp
@@ -24,26 +24,47 @@ namespace DFW
         _ecs->UpdateECSImGui();
     }
 
+    GameObject GameWorld::SpawnGameObject(SpawnInfo const& a_spawn_info)
+    {
+        return SpawnGameObject<"Entity">(a_spawn_info);
+    }
+
     void GameWorld::DestroyGameObject(GameObject const& a_game_object)
     {
+        if (!IsEntityAGameObject(a_game_object))
+        {
+            DFW_WARNLOG("Attempting to destroy a GameObject which it is not. ");
+            return;
+        }
+
         _ecs->EventHandler().Broadcast<GameObjectDestroyedEvent>(a_game_object);
 
         _ecs->EntityManager().DestroyEntity(a_game_object);
     }
 
-
     GameObject GameWorld::GetGameObject(DUID const a_game_object_id) const
     {
-        Entity entity(_ecs->EntityManager().GetEntity(a_game_object_id, _ecs->GetRegistry()));
-        if (IsGameObject(entity))
-            return GameObject(entity);
+        GameObject found_entity(_ecs->GetRegistry().GetEntity(a_game_object_id));
+        
+        if (found_entity.IsEntityValid() && IsEntityAGameObject(found_entity))
+            return found_entity;
         else
             return GameObject();
     }
 
-    bool GameWorld::IsGameObject(Entity const& a_entity) const
+    GameObject GameWorld::GetGameObject(std::string const& a_game_object_name) const
     {
-        return a_entity.TryGetComponent<GameObjectTagComponent>() ? true : false;
+        GameObject found_entity(_ecs->GetRegistry().GetEntity(a_game_object_name));
+
+        if (found_entity.IsEntityValid() && IsEntityAGameObject(found_entity))
+            return found_entity;
+        else
+            return GameObject();
+    }
+
+    bool GameWorld::IsEntityAGameObject(Entity const& a_entity) const
+    {
+        return a_entity.TryGetComponent<GameObjectTagComponent>();
     }
 
 } // End of namespace ~ DFW.
